Refactor Operator test packages
