+++ operator/app/src/test/java/org/keycloak/operator/testsuite/integration/KeycloakDeploymentTest.java
@@ -1,13 +1,32 @@
-package org.keycloak.operator;
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.keycloak.operator.testsuite.integration;
 
 import io.fabric8.kubernetes.api.model.EnvVarBuilder;
 import io.fabric8.kubernetes.api.model.apps.DeploymentSpecBuilder;
 import io.quarkus.logging.Log;
 import io.quarkus.test.junit.QuarkusTest;
+import org.assertj.core.api.Assertions;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
-import org.keycloak.operator.utils.K8sUtils;
+import org.keycloak.operator.Constants;
+import org.keycloak.operator.testsuite.utils.K8sUtils;
 import org.keycloak.operator.controllers.KeycloakAdminSecret;
 import org.keycloak.operator.controllers.KeycloakDeployment;
 import org.keycloak.operator.controllers.KeycloakService;
@@ -27,12 +46,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.keycloak.operator.utils.K8sUtils.deployKeycloak;
-import static org.keycloak.operator.utils.K8sUtils.getDefaultKeycloakDeployment;
-import static org.keycloak.operator.utils.K8sUtils.waitForKeycloakToBeReady;
+import static org.keycloak.operator.testsuite.utils.K8sUtils.deployKeycloak;
+import static org.keycloak.operator.testsuite.utils.K8sUtils.getDefaultKeycloakDeployment;
+import static org.keycloak.operator.testsuite.utils.K8sUtils.waitForKeycloakToBeReady;
 
 @QuarkusTest
-public class KeycloakDeploymentE2EIT extends ClusterOperatorTest {
+public class KeycloakDeploymentTest extends BaseOperatorTest {
     @Test
     public void testBasicKeycloakDeploymentAndDeletion() {
         try {
@@ -105,7 +124,7 @@ public class KeycloakDeploymentE2EIT extends ClusterOperatorTest {
             kc.getSpec().getServerConfiguration().add(health);
             deployKeycloak(k8sclient, kc, false);
 
-            assertThat(Constants.DEFAULT_DIST_CONFIG.get(health.getName())).isEqualTo("true"); // just a sanity check default values did not change
+            Assertions.assertThat(Constants.DEFAULT_DIST_CONFIG.get(health.getName())).isEqualTo("true"); // just a sanity check default values did not change
 
             Awaitility.await()
                     .ignoreExceptions()
